Changing /api/submit to /api/entry/{date}
This makes it more consistent with the /api/draft routes.
